feat : 코인 선택 영어 검색 기능 개발 #115
